% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject_lookup.R
\name{candidate_subjects}
\alias{candidate_subjects}
\title{Subjects worth looking up on Wikidata}
\usage{
candidate_subjects(text)
}
\arguments{
\item{text}{Character vector of subject/name text.}
}
\value{
Unique subjects X, in first-seen order, from text that starts
  with an object prefix (see [extract_subject()]); X of fewer than 3
  characters is dropped.
}
\description{
Subjects worth looking up on Wikidata
}
