% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{clean_names}
\alias{clean_names}
\title{Clean and standardize names}
\usage{
clean_names(name_vector)
}
\arguments{
\item{name_vector}{Character vector of names}
}
\value{
Cleaned character vector
}
\description{
Removes honorifics, dates, and extra whitespace from names.
}
