% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{extract_coords_from_wkt}
\alias{extract_coords_from_wkt}
\title{Extract coordinates from WKT Point}
\usage{
extract_coords_from_wkt(wkt)
}
\arguments{
\item{wkt}{Character string (e.g., "Point(-0.1 51.5)")}
}
\value{
A named vector with lat and lon
}
\description{
Extract coordinates from WKT Point
}
