% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sources.R
\name{fetch_art_uk_data}
\alias{fetch_art_uk_data}
\title{Fetch data from Art UK}
\usage{
fetch_art_uk_data(pages = 1)
}
\arguments{
\item{pages}{Integer, number of pages to scrape (default: 1)}
}
\value{
A tibble containing Art UK data
}
\description{
Scrapes statue and sculpture data from Art UK.
}
