% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sources.R
\name{fetch_historic_england_data}
\alias{fetch_historic_england_data}
\title{Fetch data from Historic England}
\usage{
fetch_historic_england_data(pages = 1)
}
\arguments{
\item{pages}{Integer, number of pages (default 1)}
}
\value{
A tibble containing Historic England data
}
\description{
Fetches listed building data for statues.
}
