% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sources.R
\name{fetch_osm_statues}
\alias{fetch_osm_statues}
\title{Fetch statues from OpenStreetMap}
\usage{
fetch_osm_statues()
}
\value{
A tibble containing OSM results
}
\description{
Queries the Overpass API for memorials and statues in London.
}
