% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sources.R
\name{fetch_wikidata_statues}
\alias{fetch_wikidata_statues}
\title{Fetch statues from Wikidata}
\usage{
fetch_wikidata_statues()
}
\value{
A tibble containing Wikidata results
}
\description{
Queries Wikidata for statues and public art in London using SPARQL.
}
