% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorials.R
\name{get_memorial}
\alias{get_memorial}
\title{Get detailed information about a specific memorial}
\usage{
get_memorial(memorial_url)
}
\arguments{
\item{memorial_url}{Character string, URL or slug of the memorial}
}
\value{
A list containing detailed memorial information
}
\description{
Get detailed information about a specific memorial
}
\examples{
\dontrun{
memorial <- get_memorial("/memorials/john-lennon-statue")
}
}
