% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorials.R
\name{get_memorials_latest}
\alias{get_memorials_latest}
\title{Get latest memorials from London Remembers}
\usage{
get_memorials_latest(pages = 1)
}
\arguments{
\item{pages}{Integer, number of pages to fetch (default: 1)}
}
\value{
A tibble containing memorial information with columns:
  \describe{
    \item{title}{Memorial title}
    \item{url}{URL to memorial page}
    \item{memorial_type}{Type of memorial (statue, plaque, etc.)}
    \item{subjects}{Subjects commemorated}
    \item{location}{Location of the memorial}
  }
}
\description{
Fetches the most recently added memorials from the London Remembers website.
}
\examples{
\dontrun{
memorials <- get_memorials_latest(pages = 2)
}
}
