% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjects.R
\name{get_subject}
\alias{get_subject}
\title{Get detailed information about a specific subject}
\usage{
get_subject(subject_url)
}
\arguments{
\item{subject_url}{Character string, URL or slug of the subject}
}
\value{
A list containing detailed subject information including associated memorials
}
\description{
Get detailed information about a specific subject
}
\examples{
\dontrun{
subject <- get_subject("/subjects/john-lennon")
}
}
