% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{is_man_named_john}
\alias{is_man_named_john}
\title{Check if a subject is a "Man named John"}
\usage{
is_man_named_john(name, gender = "male")
}
\arguments{
\item{name}{Character string of the subject's name}

\item{gender}{Character string (optional)}
}
\value{
Logical
}
\description{
Check if a subject is a "Man named John"
}
