% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{join_and_clean_data}
\alias{join_and_clean_data}
\title{Normalize and Merge Data Sources}
\usage{
join_and_clean_data(wikidata_df, osm_df)
}
\arguments{
\item{wikidata_df}{Tibble from fetch_wikidata_statues}

\item{osm_df}{Tibble from fetch_osm_statues}
}
\value{
A combined tibble
}
\description{
Combines Wikidata and OSM data into a single clean dataset.
}
