% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memorials.R
\name{search_memorials}
\alias{search_memorials}
\title{Search for memorials}
\usage{
search_memorials(query, pages = 1)
}
\arguments{
\item{query}{Character string, search term}

\item{pages}{Integer, number of pages to fetch (default: 1)}
}
\value{
A tibble containing search results
}
\description{
Search for memorials
}
\examples{
\dontrun{
results <- search_memorials("John", pages = 3)
}
}
