% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_statues.R
\name{compare_johns_vs_women}
\alias{compare_johns_vs_women}
\title{Compare John Statues vs Women Statues}
\usage{
compare_johns_vs_women(statue_data)
}
\arguments{
\item{statue_data}{Standardized statue data tibble}
}
\value{
A list with comparison results:
  - total_statues, john_statues, woman_statues, john_percent, woman_percent,
    claim_validated, message (as before)
  - unknown_statues, unknown_percent: statues whose gender could not be
    confidently classified (see classify_gender_from_subject())
  - gender_method: which classification sources were used, in priority
    order (Wikidata P21, then the genderdata lookup cascade)
}
\description{
Validates the "Statues for Equality" claim that there are more statues
named John than women in the UK.
}
